GDB/MI grammar update - result and value merged.
When adding the test for out-of-band records that have async_record objects
of kind GDBMI_STATUS, I discovered the flaw in the GDB/MI specification
regarding results and values. GDB produces output that the specification can
not parse. This fixes the parser to handle GDB's real output and documents
the changes in gdbmi_grammar.txt.

This also adds the new test, oob_record/async/status/basic in gdbmi_pt.cpp
which was responsible for adding the new enumeration value
GDBMI_ASYNC_DOWNLOAD to the enumeration gdbmi_async_class.