Improve support for gdbmi_token_t.
Previously, was supporting the token part of the grammar rule with an
integer. This was wrong. GDB supports this functionality with a string.
So 0000 is valid for instance.

Added the unit test result_record/token to support the result records
token field. Documented that async records never have a token populated.