Allow unit tests to refer to a file or directory.
Previously, if deriving a unit test off of the Fixture class
in fixture.h unit tests could use the TEST_CASE_METHOD_N macro
which expected a directory name in the second argument.

Now the second argument can be either a file or a directory.

This was useful in the gdbmi unit tests so that the directory
structure containing the unit tests could have more than one
mi file per directory. Previously it was limited to 1 per
directory which made adding new tests slightly awkward.

Previous:
  oob_record/async/exec/async_class (with input.mi in directory)
  oob_record/async/exec/basic (with input.mi in directory)
Now:
  oob_record/async/exec/async_class.mi
  oob_record/async/exec/basic.mi

This will be particularly useful as unit tests are added next
for the gdbmi_result struct in gdbmi. This will require lots
of small unit tests to cover the edge cases and I was not
excited about the idea of spreading those input mi files over
lots of directories.