Remove the ability to print the parse tree.
This functionality has no purpose in the library at this point.
I thought at one point it would be useful for unit testing but
I decided against using it for this purpose.

Then I thought it might be useful for debugging. However, now
it's easy to print a GDB/MI line and the token at which it is an error.

Dropping support for this functionality until it becomes necessary again.