The parser is now sometimes much faster.
Before this patch, if the user passed in a single character at a
time into the parser, the parser would traverse the entire string
it has received so far looking for a new line.

Now it only looks at the input data once.

After adding the large example, and improving the runtime perforance,
the test suite still jumped up .5 seconds. I modified the way we
read files in the test suite to read the entire file, instead of char
by char. The test suite is now fast again, however, we are no longer
directly testing the char-by-char case that was initially found
to be slow.