Fix -file-list-exec-source-files parsing bug.
Previously, if a -file-list-exec-source-files response had the field,
debug-fully-read (which was introduced in gdb 11), then gdbwire
would fail in file_list_exec_source_files() and return an error.
Now it will successfuly parse the gdb response.

In addition, the gdbwire_mi_source_file structure now has a new field,
debug_fully_read, which tracks the state of the gdb response.

Signed-off-by: Bob Rossi <bob@brasko.net>